requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.15
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
joblib==1.4.2
numpy==1.24.4
PyJWT==2.9.0
scikit-learn==1.3.2
scipy==1.10.1
sqlparse==0.5.1
threadpoolctl==3.5.0
typing-extensions==4.12.2